Moved controller stuff out of the modpack selector files for cleanliness
